Fixed UIUtil.setBackgroundCoverForView to check for write permissions
Falls back to original file if write permissions are missing on Android. This is needed because of a bug in how Titanium handles storage permissions: applicationCacheDirectory and other directories shouldn't need storage permissions, but they do. This will be fixed in Titanium 5.4.0. Maybe. See: https://jira.appcelerator.org/browse/TIMOB-20440