Refactored SQLite again. Added SQLite.run() and SQLite.transform().
SQLite.execute() now has the previous behaviour, so it will keep the database and returned dataset open. SQLite.run() will simply execute a query chain or a query and close the dataset and the database immediately after. SQLite.transform() accepts a query and calls a transformation function with the ResultSet as parameter. See the docs for more details.